fix: 主干告警清零 【背景】 主干中存在告警需要处理
【修改方案】
在有符号数与无符号数混用的地方使用了强制类型转换

【影响】
对现有的产品编译不会有影响。

re #I4SPC0

+++ kal/posix/src/mqueue.c
@@ -384,7 +384,7 @@ int OsMqGetAttr(mqd_t personal, struct mq_attr *mqAttr)
 
     mqueueCB = privateMqPersonal->mq_posixdes;
     mqAttr->mq_maxmsg = mqueueCB->mqcb->queueLen;
-    mqAttr->mq_msgsize = mqueueCB->mqcb->queueSize - sizeof(UINT32);
+    mqAttr->mq_msgsize = (long)(mqueueCB->mqcb->queueSize - sizeof(UINT32));
     mqAttr->mq_curmsgs = mqueueCB->mqcb->readWriteableCnt[OS_QUEUE_READ];
     mqAttr->mq_flags = privateMqPersonal->mq_flags;
     (VOID)pthread_mutex_unlock(&g_mqueueMutex);
